hallo

+++ main.py
@@ -1,29 +1,29 @@
 from selenium import webdriver
 from selenium.webdriver.common.by import By
-from selenium.webdriver.common.keys import Keys
-import time
 from datetime import datetime
 import json
-
+from os.path import exists
 
 if __name__ == '__main__':
 
+    today = str(datetime.now()).split('.')[0]
     print(datetime.now())
 
     datas = {}
     loop = True
     counter = 1
+    count = 0
 
     browser = webdriver.Chrome()
     new_browser = webdriver.Chrome()
     browser.get('https://www.pflegeausbildung.net/no_cache/alles-zur-ausbildung/uebersicht-pflegeschulen.html')
 
-    while loop:
-        print(len(datas))
+    # while loop:
+    while count == 0:
+        count += 1
         list_schools = browser.find_element(By.CLASS_NAME, 'altenpflegeschulen')
         singleItems = list_schools.find_elements(By.CLASS_NAME, 'showSingleItem')
         links = [item.get_attribute('href') for item in singleItems]
-
         for item in links:
             new_browser.get(item)
             details = new_browser.find_element(By.CLASS_NAME, 'detailView')
@@ -38,6 +38,9 @@ if __name__ == '__main__':
                 if '<p>' in content.get_attribute('innerHTML'):
                     p = content.find_element(By.TAG_NAME, 'p')
                     infos = (p.get_attribute('innerText').split('\n'))
+                    street = infos[1]
+                    plz = infos[2].split(' ')[0]
+                    city = infos[2].split(' ')[1]
                 else:
                     children = (content.find_element(By.TAG_NAME, 'ul').get_property('children'))
                     for child in children:
@@ -64,8 +67,9 @@ if __name__ == '__main__':
 
             data = {
                 'Name': details.find_element(By.TAG_NAME, 'h2').text,
-                'Straße': infos[1],
-                'PLZ/Ort': infos[2],
+                'Straße': street,
+                'PLZ': plz,
+                'Stadt': city,
                 'Telefon': telefon,
                 'EMail': email,
                 'Web': web,
@@ -83,8 +87,23 @@ if __name__ == '__main__':
 
     print('##########################################')
     print(len(datas))
-    with open('data.json', 'w', encoding='utf-8') as f:
-        json.dump(datas, f, ensure_ascii=False, indent=4)
+    if exists('data.json'):
+        with open('data.json', 'r', encoding='utf-8') as f:
+            content = json.load(f)
+            for data in datas:
+                for key in content:
+                    if datas[data]['Name'] == content[key]['Name']:
+                        for data_key in datas[data]:
+                            content[key][data_key] = datas[data][data_key]
+                        content[data]['Überprüft'] = today
+        with open('data.json', 'w', encoding='utf-8') as nf:
+            json.dump(content, nf, ensure_ascii=False, indent=4)
+    else:
+        for data in datas:
+            datas[data]['Eintrag'] = today
+            datas[data]['Überprüft'] = today
+        with open('data.json', 'w', encoding='utf-8') as f:
+            json.dump(datas, f, ensure_ascii=False, indent=4)
     print('##########################################')
     print(datetime.now())
